requirements.txt
asgiref==3.10.0
attrs==25.4.0
black==25.11.0
click==8.3.1
Django==5.2.8
django-debug-toolbar==6.1.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
psycopg==3.2.13
psycopg-binary==3.2.13
PyJWT==2.10.1
pytokens==0.3.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
sqlparse==0.5.3
uritemplate==4.2.0
